Split c++ targets of core_worker
