+++ src/ray/core_worker/store_provider/memory_store/memory_store.cc
@@ -22,7 +22,6 @@
 
 #include "ray/common/ray_config.h"
 #include "ray/core_worker/context.h"
-#include "ray/core_worker/core_worker.h"
 
 namespace ray {
 namespace core {
